requirements.txt
langchain
langchain_google_genai
streamlit
python-dotenv
